refactor: extract normalizePlaceholders, formatJson, formatTime to shared utils
